测试mysql-connector-c/6.1.11

+++ billing/inc/server_config.hpp
@@ -9,8 +9,8 @@ public:
 	ServerConfig();
 	~ServerConfig();
 	const string& getIp();
-	const int& getPort();
+	const unsigned short& getPort();
 private:
 	string ip;
-	int port;
+	unsigned short port;
 };
